Serialize DateTimes in ISO8061 format in JSON responses
It seems that setting *any* (non-null) date format on the object mapper causes DateTimes to be serialized in ISO8061 format, i.e.

    static {
        mapper.setDateFormat(new SimpleDateFormat(""));
    }

would cause DateTime objects to be serialized in ISO8061 format too. This appears to be a bug in Jackson version 1.9.10.

I set the format to "yyyy-MM-dd'T'HH:mm:ssz" to try to make it clearer why a date format is being set on the ObjectMapper.