Allow searching in a thirdparty folder for vendor jar files.
This means that the maven copy-dependencies jar-with-dependencies goal have its
output entirely placed into the `thirdparty` folder, making deployment easier.
This stops conflicts with the existing oncue-agent dependencies in the `lib`
folder of the microkernel.

Note: This may cause issues if the thirdparty libraries depend on a different
version of the oncue-agent core libraries (e.g. logback), as there will be
multiple versions of some classes on the classpath. This will also be an issue
with placing your dependencies in the `lib` folder currently.